GuitarHeroVisualizer.
This is intended to graph the wave as the keys are played and the
energy of each vibration. It is currently static and the wave forms do
not change to match the energy levels of the vibrations. Also, when
some keys are played, a line is being drawn across the screen, which
should not be there.